% Generated by roxygen2 (4.0.1): do not edit by hand
\name{contLikMarg}
\alias{contLikMarg}
\title{contLikMarg}
\usage{
contLikMarg(nC, mixData, popFreq, lower, upper, refData = NULL,
  condOrder = NULL, knownRef = NULL, xi = NULL, prC = 0,
  reltol = 0.001, threshT = 50, fst = 0, lambda = 0, pXi = function(x)
  1, isPhi = FALSE)
}
\arguments{
\item{nC}{Number of contributors in model.}

\item{mixData}{Evidence object with list elements adata[[i]] and hdata[[i]]. Each element has a loci-list with list-element 'i' storing qualitative data in 'adata' and quantitative data in 'hdata'.}

\item{popFreq}{A list of allele frequencies for a given population.}

\item{lower}{Lower bounds of parameters}

\item{upper}{Upper bounds of parameters}

\item{refData}{Reference objects with list element [[s]]$adata[[i]]. The list element has reference-list with list-element 's' having a loci-list adata with list-element 'i storing qualitative data.}

\item{condOrder}{Specify conditioning references from refData (must be consistent order). For instance condOrder=(0,2,1,0) means that we restrict the model such that Ref2 and Ref3 are respectively conditioned as 2. contributor and 1. contributor in the model. condOrder=-1 means the reference is known-non contributor!}

\item{knownRef}{Specify known references from refData (index). For instance knownRef=(1,2) means that reference 1 and 2 is known allele samples in the hypothesis. This is affected by fst-correction.}

\item{xi}{A numeric giving stutter-ratio if it is known. Default is NULL, meaning it is integrated out.}

\item{prC}{A numeric for allele drop-in probability. Default is 0.}

\item{reltol}{Required relative tolerance error of evaluations in integration routine. Default is 0.001.}

\item{threshT}{The detection threshold given. Used when considering probability of allele drop-outs.}

\item{fst}{is the coancestry coeffecient. Default is 0.}

\item{lambda}{Parameter in modeled peak height shifted exponential model. Default is 0.}

\item{pXi}{Prior function for xi-parameter (stutter). Flat prior on [0,1] is default.}

\item{isPhi}{Boolean whether we are working with reparameterizated parameters}
}
\value{
ret A list(margL,deviation,nEvals) where margL is Marginalized likelihood for hypothesis (model) given observed evidence, deviation is the confidence-interval of margL, nEvals is number of evaluations.
}
\description{
contLikMarg marginalized the likelihood of the STR DNA mixture given some assumed a bayesian model.
}
\details{
The procedure are doing numerical integration to approximate the marginal probability over noisance parameters. Mixture proportions have flat prior.

Function calls procedure in c++ by using the package Armadillo and Boost.
Using "cubature" method (cubature) or "divonne" method (R2Cuba). "cubature" gives more accuracy but require more evaluations than "divonne".
"cuhre","suave" and "vegas" not recommended by experience.
}
\examples{
\dontrun{
threshT <- 50 #threshold
load(system.file("mcData.Rdata", package = "gammadnamix"))
mixData <- data$samples$MC15
retlist <- Qassignate(mixData,popFreq=data$popFreq,refData=data$refData) #Q-assignation
popFreq <- retlist$popFreq #get updated popFreq
refData <- retlist$refData #get updated popFreq
#Hypothesis:
nC <- 2 #number of contributors
condhp <- condhd <- rep(0,length(refData))
condhp [1] <- 1 #condition on first one in refdata
#without stutter:
hpD <- contLikMarg(nC,mixData,popFreq,refData=refData,condOrder=condhp,xi=0,threshT=threshT)
hdD <- contLikMarg(nC,mixData,popFreq,refData=refData,condOrder=condhd,xi=0,threshT=threshT)
print(hpD$margL/hdD$margL) #estimated LR=8670.289
#with stutter:
hpS <- contLikMarg(nC,mixData,popFreq,refData=refData,condOrder=condhp,threshT=threshT,ximax=0.2)
hdS <- contLikMarg(nC,mixData,popFreq,refData=refData,condOrder=condhd,threshT=threshT,ximax=0.2)
print(hpS$margL/hdS$margL) #estimated LR=8645.057
}
}
\author{
Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}
\references{
Hahn,T. (2005). CUBA - a library for multidimensional numerical integration. Computer Physics Communications, 168(2),78-95.
}
\keyword{Bayesian}
\keyword{Likelihood}
\keyword{Marginalized}
\keyword{continuous,}
\keyword{estimation}
\keyword{models,}

