% Generated by roxygen2 (4.0.1): do not edit by hand
\name{contLikMCMC}
\alias{contLikMCMC}
\title{contLikMCMC}
\usage{
contLikMCMC(nC, mixData, popFreq, refData = NULL, condOrder = NULL,
  xi = NULL, prC = 0, method = 2, pRhoTau = function(x) {     1 },
  rhotaumax = c(1000, 1000), nDone = 1, threshT = 50, fst = 0,
  lambda = 0, pXi = function(x) {     1 }, ximax = 1, M = 10000,
  theta0 = NULL, Sigma = NULL, delta = 0.3)
}
\arguments{
\item{nC}{Number of contributors in model.}

\item{mixData}{Evidence object with list elements adata[[i]] and hdata[[i]]. Each element has a loci-list with list-element 'i' storing qualitative data in 'adata' and quantitative data in 'hdata'.}

\item{popFreq}{A list of allele frequencies for a given population.}

\item{refData}{Reference objects with list element [[s]]$adata[[i]]. The list element has reference-list with list-element 's' having a loci-list adata with list-element 'i storing qualitative data.}

\item{condOrder}{Specify conditioning references from refData (must be consistent order). For instance condOrder=(0,2,1,0) means that we restrict the model such that Ref2 and Ref3 are respectively conditioned as 2. contributor and 1. contributor in the model. condOrder=-1 means the reference is known-non contributor!}

\item{xi}{A numeric giving stutter-ratio if it is known. Default is NULL, meaning it is integrated out.}

\item{prC}{A numeric for allele drop-in probability. Default is 0.}

\item{method}{Selected MCMC-routine for calculate marginal Likelihood (1=Normal Approximation, 2=Metropolis Hastings).}

\item{pRhoTau}{Prior function for (rho,tau)-parameters. Flat prior is default.}

\item{rhotaumax}{Maximum range of rho and tau-parameter. Default is c(1000,1000).}

\item{nDone}{Maxumum number of random evaluations nlm-optimizing routing. Default is 1.}

\item{threshT}{The detection threshold given. Used when considering probability of allele drop-outs.}

\item{fst}{is the coancestry coeffecient. Default is 0.}

\item{lambda}{Parameter in modeled peak height shifted exponential model. Default is 0.}

\item{pXi}{Prior function for xi-parameter (stutter). Flat prior on [0,1] is default.}

\item{ximax}{Maximum range of xi-parameter. Default is 1.}

\item{M}{Number of evaluations for MCMC-sampling.}

\item{theta0}{Start values for parameters in MCMC-sampling. Default is to find these using MLE.}

\item{Sigma}{Covariance matrix used in kernel-function in MCMC-sampling. Default is to use inverse -hessian from optimization. If theta0 is spesified but not Sigma, Sigma becomes an eye-matrix.}

\item{delta}{A numerical parameter to scale with Sigma. Useful for better MCMC sampling.}
}
\value{
ret A list(margL,posttheta,postlogL,logpX,nacc) where margL is Marginalized likelihood for hypothesis (model) given observed evidence, posttheta is the posterior samples from a MC routine, postlogL is sampled log-likelihood values, nacc is number of accepted samples.
}
\description{
contLikMCMC simulates from the posterior distribution for a bayesian STR DNA mixture model.
}
\details{
The procedure are doing MCMC to approximate the marginal probability over noisance parameters. Mixture proportions have flat prior.

Function calls procedure in c++ by using the package Armadillo and Boost.
}
\examples{
\dontrun{
#data model:
threshT <- 50 #threshold
load(system.file("mcData.Rdata", package = "gammadnamix"))
popFreq <- data$popFreq
mixData <- data$samples$MC15
refData <- data$refData
#Q-assignation
popFreq2 <- popFreq
for(i in 1:length(popFreq)) { #make Q-assignation for each loci
 tmp <- popFreq[[i]][names(popFreq[[i]])\%in\%mixData$adata[[i]]]
 tmp <- c(tmp,1-sum(tmp))
 names(tmp)[length(tmp)] <- "99"
 popFreq2[[i]] <- tmp
}
#Hypothesis:
nC <- 2 #number of contributors
condhp <- condhd <- rep(0,length(refData))
condhp [1] <- 1 #condition on first one in refdata
#helpfunction for checking samples:
validMCMC <- function(X) {
 p <- ncol(X)
 par(mfrow=c(p,2))
 for(i in 1:p) {
  plot(density(X[,i],from=0),ty="l")
  plot(X[,i],ty="l")
 }
 par(mfrow=c(1,1))
}
M <- 1e5 #number of samples from posterior
delta = 0.3 #tweak parameter in MCMC
#without stutter:
hpD4 <- contLikMCMC(nC,mixData,popFreq2,refData=refData,condOrder=condhp,xi=0,threshT=threshT,M=M,delta=delta)
hdD4 <- contLikMCMC(nC,mixData,popFreq2,refData=refData,condOrder=condhd,xi=0,threshT=threshT,M=M,delta=delta)
validMCMC(hpD4$posttheta)
validMCMC(hpD4$posttheta)
LRD4 <- hpD4$margL/hdD4$margL #estimated LR
#with stutter:
hpS4 <- contLikMCMC(nC,mixData,popFreq2,refData=refData,condOrder=condhp,threshT=threshT,M=M,ximax=1,delta=delta)
hdS4 <- contLikMCMC(nC,mixData,popFreq2,refData=refData,condOrder=condhd,threshT=threshT,M=M,ximax=1,delta=delta)
validMCMC(hpS4$posttheta)
validMCMC(hdS4$posttheta)
LRS4 <- hpS4$margL/hdS4$margL #estimated LR
}
}
\author{
Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}
\keyword{Bayesian}
\keyword{continuous,}
\keyword{models}

