% Generated by roxygen2 (4.0.1): do not edit by hand
\name{contLikMLE}
\alias{contLikMLE}
\title{contLikMLE}
\usage{
contLikMLE(nC, mixData, popFreq, refData = NULL, condOrder = NULL,
  xi = NULL, prC = 0, pRhoTau = function(x) {     1 },
  rhotaumax = c(1000, 1000), nDone = 1, threshT = 50, fst = 0,
  lambda = 0, pXi = function(x) {     1 }, M = NULL)
}
\arguments{
\item{nC}{Number of contributors in model.}

\item{mixData}{Evidence object with list elements adata[[i]] and hdata[[i]]. Each element has a loci-list with list-element 'i' storing qualitative data in 'adata' and quantitative data in 'hdata'.}

\item{popFreq}{A list of allele frequencies for a given population.}

\item{refData}{Reference objects with list element [[s]]$adata[[i]]. The list element has reference-list with list-element 's' having a loci-list adata with list-element 'i storing qualitative data.}

\item{condOrder}{Specify conditioning references from refData (must be consistent order). For instance condOrder=(0,2,1,0) means that we restrict the model such that Ref2 and Ref3 are respectively conditioned as 2. contributor and 1. contributor in the model. condOrder=-1 means the reference is known-non contributor!}

\item{xi}{A numeric giving stutter-ratio if it is known. Default is NULL, meaning it is integrated out.}

\item{prC}{A numeric for allele drop-in probability. Default is 0.}

\item{pRhoTau}{Prior function for (rho,tau)-parameters. Flat prior is default.}

\item{rhotaumax}{Maximum range of rho and tau-parameter. Default is c(1000,1000).}

\item{nDone}{Maxumum number of random evaluations nlm-optimizing routing. Default is 1.}

\item{threshT}{The detection threshold given. Used when considering probability of allele drop-outs.}

\item{fst}{is the coancestry coeffecient. Default is 0.}

\item{lambda}{Parameter in modeled peak height shifted exponential model. Default is 0.}

\item{pXi}{Prior function for xi-parameter (stutter). Flat prior on [0,1] is default.}

\item{M}{Number of evaluations for importance sampling.}
}
\value{
ret A list(MLE,Sigma,CI,loglik,margL) with Maximixed likelihood elements for hypothesis (model) given observed evidence.
}
\description{
contLikMLE optimizes the likelihood of the STR DNA mixture given some assumed a bayesian model.
}
\details{
The procedure are doing numerical optimization to approximate the marginal probabilit over noisance parameters. Mixture proportions have flat prior.

Function calls procedure in c++ by using the package Armadillo and Boost.
}
\author{
Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}
\keyword{Bayesian}
\keyword{continuous,}
\keyword{models}

