% Generated by roxygen2 (4.0.1): do not edit by hand
\name{deconvolve}
\alias{deconvolve}
\title{deconvolve}
\usage{
deconvolve(mlefit, alpha = 0.95, maxlist = 1000)
}
\arguments{
\item{mlefit}{Fitted object using contLikMLE function.}

\item{alpha}{Required sum of the listed posterior probabilities.}

\item{maxlist}{The ranked deconvolved profile list will not exceed this number (used to avoid endless search).}
}
\value{
ret A list(table1,rankG,pG) where rankG is the ranked genotypes with corresponding probabilities in pG. table1 is a formated version of these two.
}
\description{
deconvolve ranks the set of the most conditional posterior probability of genotypes the STR DNA mixture given a fitted model under a hypothesis.
}
\details{
The procedure calculates the likelihood for each single locus. Then it combines the most probable genotypes from each loci to produce a ranked list of deconvolved profiles.

Function calls procedure in c++ by using the package Armadillo and Boost.
}
\author{
Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}
\references{
Cowell,R.G. et.al. (2014). Analysis of forensic DNA mixtures with artefacts. Applied Statistics, 64(1),1-32.
}
\keyword{deconvolution}

