% Generated by roxygen2 (4.0.1): do not edit by hand
\name{euroformix}
\alias{euroformix}
\title{euroformixTK}
\usage{
euroformix(envirfile = NULL)
}
\arguments{
\item{envirfile}{A file to a saved environment of a project}
}
\description{
euroformixTK is a GUI wrapper for gammadnamix
}
\details{
The function is a graphical layer for the functions in the package gammadnamix. See ?gammadnamix for more information.
}
\author{
Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}

