% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotEPG}
\alias{plotEPG}
\title{plotEPG}
\usage{
plotEPG(Data, kitname, sname = "")
}
\arguments{
\item{Data}{List of adata- and hdata-elements.}

\item{kit}{name of kit: {"ESX17","ESI17","ESI17Fast","ESX17Fast","Y23","Identifiler","NGM","ESSPlex","ESSplexSE","NGMSElect","SGMPlus","ESX16", "Fusion","GlobalFiler"}}

\item{sname}{Sample name label.}
}
\description{
EPG plotter created by Oskar Hansson.
}
\details{
Plots peak height with corresponding allele for one sample for a given kit.
}
\author{
Oskar Hansson <Oskar.Hansson.at.fhi.no> and Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}

