% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prepareC}
\alias{prepareC}
\title{prepareC}
\usage{
prepareC(nC, mixData, popFreq, refData = NULL, condOrder = NULL,
  knownRef = NULL)
}
\arguments{
\item{nC}{Number of contributors in model.}

\item{mixData}{Evidence object with list elements adata[[i]] and hdata[[i]]. Each element has a loci-list with list-element 'i' storing qualitative data in 'adata' and quantitative data in 'hdata'.}

\item{popFreq}{A list of allele frequencies for a given population.}

\item{refData}{Reference objects with list element [[s]]$adata[[i]]. The list element has reference-list with list-element 's' having a loci-list adata with list-element 'i storing qualitative data.}

\item{condOrder}{Specify conditioning references from refData (must be consistent order). For instance condOrder=(0,2,1,0) means that we restrict the model such that Ref2 and Ref3 are respectively conditioned as 2. contributor and 1. contributor in the model. condOrder=-1 means the reference is known-non contributor!}

\item{knownRef}{Specify known references from refData (index). For instance knownRef=(1,2) means that reference 1 and 2 is known allele samples in the hypothesis. This is affected by fst-correction.}
}
\value{
ret A list of data input to call the C-code with
}
\description{
prepareC is used in the functions contLikMLE,contLikMarg and contLikMCMC to prepare input to C-call
}
\details{
The function builds the data input to the C-code
}
\author{
Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}

