% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Qassignate}
\alias{Qassignate}
\title{Qassignate}
\usage{
Qassignate(mixData, popFreq, refData = NULL)
}
\arguments{
\item{mixData}{Evidence object with list elements adata[[i]] and hdata[[i]]. Each element has a loci-list with list-element 'i' storing qualitative data in 'adata' and quantitative data in 'hdata'.}

\item{popFreq}{A list of allele frequencies for a given population.}

\item{refData}{Reference objects with list element [[s]]$adata[[i]]. The list element has reference-list with list-element 's' having a loci-list adata with list-element 'i storing qualitative data.}
}
\value{
ret A list(popFreq,refData) with Q-assignated alleles.
}
\description{
Q-assignation
}
\details{
Assignes non-shown alleles as one single allele.
}
\author{
Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}

