% Generated by roxygen2 (4.0.1): do not edit by hand
\name{validMCMC}
\alias{validMCMC}
\title{validMCMC}
\usage{
validMCMC(X, trace = TRUE, acf = TRUE, mlefit = NULL, isPhi = FALSE)
}
\arguments{
\item{X}{A(Mxp) matrix with MCMC samples.}

\item{trace}{Boolean for whether showing trace of samples.}

\item{acf}{Boolean for whether showing autocorrelation function of samples.}

\item{mlefit}{Fitted object using contLikMLE}

\item{isPhi}{Boolean whether we are working with reparameterizited parameters}
}
\description{
Validates aposteriori samples from MCMC method
}
\details{
This function takes samples from the MCMC as given in a matrix and shows the aposterior functions.
}
\author{
Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}

