% Generated by roxygen2 (4.0.1): do not edit by hand
\name{validMLEmodel}
\alias{validMLEmodel}
\title{validMLEmodel}
\usage{
validMLEmodel(mlefit, plotYdistr = FALSE, plotDdistr = FALSE)
}
\arguments{
\item{mlefit}{Fitted object using contLikMLE}

\item{plotYdistr}{A boolean whether to plot the model fitted peak height density for each contributors}

\item{plotDdistr}{A boolean whether to plot the model fitted heterozygote dropout probabilities as a function of amount of DNA(mu) for each contributors}
}
\value{
ret A vector for each marker with cumulative probabilities
}
\description{
validMLEmodel makes model validation whether the observed peak heights fits the maximum likelihood fitted gamma distribution.
}
\details{
The cumulative probability of the observed allele peaks are calculated and compared with a uniform distribution.
Function calls procedure in c++ by using the package Armadillo and Boost.
}
\author{
Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}

