% Generated by roxygen2 (4.0.1): do not edit by hand
\name{contLikMCMC}
\alias{contLikMCMC}
\title{contLikMCMC}
\usage{
contLikMCMC(mlefit, musigmamax = c(10000, 1), M = 10000, delta = 2,
  isPhi = FALSE)
}
\arguments{
\item{mlefit}{Fitted object using contLikMLE}

\item{method}{Selected MCMC-routine for calculate marginal Likelihood (1=Importance sampling with normal approximation, 2=.}

\item{musigmamax}{Upper boundary of rho-tau-parameters. Default (10000,1)}

\item{M}{Number of samples in the MCMC-sampling.}

\item{delta}{A numerical parameter to scale with the covariance function Sigma. Default is 2. Should be higher to obtain lower acception rate.}

\item{isPhi}{Boolean whether we are working with reparameterizated parameters}
}
\value{
ret A list (margL,posttheta,postlogL,logpX,accrat) where margL is Marginalized likelihood for hypothesis (model) given observed evidence, posttheta is the posterior samples from a MC routine, postlogL is sampled log-likelihood values, accrat is ratio of accepted samples.
}
\description{
contLikMCMC simulates from the posterior distribution for a bayesian STR DNA mixture model.
}
\details{
The procedure are doing MCMC to approximate the marginal probability over noisance parameters. Mixture proportions have flat prior.

If no initial values or covariance matrix has been provided to the function, a call to the MLE function is applied.
The Metropolis Hastings routine uses a Multivariate Normal distribution with mean 0 and covariance as delta multiplied with the inverse negative hessian with MLE inserted as transistion kernel.
Function calls procedure in c++ by using the package Armadillo and Boost.
Marginalized likelihood is estimated using Metropolis Hastings with the "Gelfand and Dey" method.
}
\author{
Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}
\references{
Craiu,R.V. and Rosenthal, J.S. (2014). Bayesian Computation Via Markov Chain Monte Carlo. Annu. Rev. Stat. Appl., 1,179-201.
}
\keyword{BayesianModels,}
\keyword{MCMC,}
\keyword{MarginalizedLikelihoodEstimation}
\keyword{MetropolisHastings,}
\keyword{continuous,}

