% Generated by roxygen2 (4.0.1): do not edit by hand
\name{contLikMLE}
\alias{contLikMLE}
\title{contLikMLE}
\usage{
contLikMLE(nC, mixData, popFreq, refData = NULL, condOrder = NULL,
  knownRef = NULL, xi = NULL, prC = 0, musigmamax = c(10000, 10),
  nDone = 1, threshT = 50, fst = 0, lambda = 0, pXi = function(x) 1,
  ximax = 1)
}
\arguments{
\item{nC}{Number of contributors in model.}

\item{mixData}{Evidence object with list elements adata[[i]] and hdata[[i]]. Each element has a loci-list with list-element 'i' storing qualitative data in 'adata' and quantitative data in 'hdata'.}

\item{popFreq}{A list of allele frequencies for a given population.}

\item{refData}{Reference objects with list element [[s]]$adata[[i]]. The list element has reference-list with list-element 's' having a loci-list adata with list-element 'i storing qualitative data.}

\item{condOrder}{Specify conditioning references from refData (must be consistent order). For instance condOrder=(0,2,1,0) means that we restrict the model such that Ref2 and Ref3 are respectively conditioned as 2. contributor and 1. contributor in the model. condOrder=-1 means the reference is known-non contributor!}

\item{knownRef}{Specify known references from refData (index). For instance knownRef=(1,2) means that reference 1 and 2 is known allele samples in the hypothesis. This is affected by fst-correction.}

\item{xi}{A numeric giving stutter-ratio if it is known. Default is NULL, meaning it is integrated out.}

\item{prC}{A numeric for allele drop-in probability. Default is 0.}

\item{musigmamax}{Maximum range of mu and sigma-parameter. Default is c(10000,10).}

\item{nDone}{Maxumum number of random evaluations nlm-optimizing routing. Default is 1.}

\item{threshT}{The detection threshold given. Used when considering probability of allele drop-outs.}

\item{fst}{is the coancestry coeffecient. Default is 0.}

\item{lambda}{Parameter in modeled peak height shifted exponential model. Default is 0.}

\item{pXi}{Prior function for xi-parameter (stutter). Flat prior on [0,1] is default.}

\item{ximax}{Maximum range of xi-parameter. Default is 1.}
}
\value{
ret A list(MLE,Sigma,Sigma2,CI,loglik) with Maximixed likelihood elements for hypothesis (model) given observed evidence.
}
\description{
contLikMLE optimizes the likelihood of the STR DNA mixture given some assumed a bayesian model.
}
\details{
The procedure are doing numerical optimization to approximate the marginal probabilit over noisance parameters. Mixture proportions have flat prior.

Function calls procedure in c++ by using the package Armadillo and Boost.
}
\examples{
\dontrun{
threshT <- 50 #threshold
load(system.file("mcData.Rdata", package = "gammadnamix"))
mixData <- data$samples$MC15
retlist <- Qassignate(mixData,popFreq=data$popFreq,refData=data$refData) #Q-assignation
popFreq <- retlist$popFreq #get updated popFreq
refData <- retlist$refData #get updated popFreq
#Hypothesis:
nC <- 2 #number of contributors
condhp <- condhd <- rep(0,length(refData))
condhp [1] <- 1 #condition on first one in refdata
nDone <- 3 #require 3 optimizations
hpD <- contLikMLE(nC,mixData,popFreq,refData=refData,condOrder=condhp,xi=0,threshT=threshT,nDone=nDone )
hdD <- contLikMLE(nC,mixData,popFreq,refData=refData,condOrder=condhd,xi=0,threshT=threshT,nDone=nDone )
print(exp(hpD$loglik - hdD$loglik)) #estimated LR=21554.24
#with stutter, but stuttarratio at most 0.2:
hpS <- contLikMLE(nC,mixData,popFreq,refData=refData,condOrder=condhp,threshT=threshT,nDone=nDone,ximax=0.2)
hdS <- contLikMLE(nC,mixData,popFreq,refData=refData,condOrder=condhd,threshT=threshT,nDone=nDone,ximax=0.2)
print(exp(hpS$loglik - hdS$loglik)) #estimated LR=21554.4
}
}
\author{
Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}
\references{
Cowell,R.G. et.al. (2014). Analysis of forensic DNA mixtures with artefacts. Applied Statistics, 64(1),1-32.
}
\keyword{Estimation}
\keyword{Likelihood}
\keyword{Maximum}
\keyword{continuous}
\keyword{model,}

