% Generated by roxygen2 (4.0.1): do not edit by hand
\name{genDataset}
\alias{genDataset}
\title{genDataset}
\usage{
genDataset(nC, popFreq, mu = 1000, sd = 1000, sorted = FALSE,
  thresh = 50)
}
\arguments{
\item{nC}{Number of contributors in model.}

\item{popFreq}{A list of allele frequencies for a given population.}

\item{mu}{Expected peak heights for an allele of a contributor.}

\item{sd}{Standard deviation of peak heights for an allele of a contributor.}

\item{sorted}{Boolean for wheter sorting the contributors with respect to decreasingly mixture proportions.}

\item{thresh}{Threshold given}
}
\value{
List with elements Mx,mixData,refData
}
\description{
genDataset samples a random mixture with Normal[+] peak heights.
}
\details{
Simple function for generating a random peak height mixture distributed as normal positive truncated.
}
\author{
Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}

