% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Qassignate}
\alias{Qassignate}
\title{Qassignate}
\usage{
Qassignate(samples, popFreq, refData = NULL)
}
\arguments{
\item{samples}{A List with samples which for each samples has locus-list elements with list elements adata and hdata. 'adata' is a qualitative (allele) data vector and 'hdata' is a quantitative (peak heights) data vector.}

\item{popFreq}{A list of allele frequencies for a given population.}

\item{refData}{Reference objects with list element [[s]]$adata[[i]]. The list element has reference-list with list-element 's' having a loci-list adata with list-element 'i storing qualitative data.}
}
\value{
ret A list(popFreq,refData) with Q-assignated alleles.
}
\description{
Q-assignation. It also takes care of include new alleles into popFreq by assigning it lowest observed frequnce.
}
\details{
Assignes non-shown alleles as one single allele.
}
\author{
Oyvind Bleka <Oyvind.Bleka.at.fhi.no>
}

